resmgr: stop importing kubernetes/kubelet internals.
Don't import kubernetes/kubelet for namespace and label
constants or milli-CPU conversion constants and functions.
Instead, define our own constants and functions for those
few things we need.

+++ pkg/cri/resource-manager/policy/builtin/balloons/balloons-policy.go
@@ -1278,7 +1278,7 @@ func (p *balloons) pinCpuMem(c cache.Container, cpus cpuset.CPUSet, mems idset.I
 		c.SetCpusetCpus(cpus.String())
 		if reqCpu, ok := c.GetResourceRequirements().Requests[corev1.ResourceCPU]; ok {
 			mCpu := int(reqCpu.MilliValue())
-			c.SetCPUShares(int64(cache.MilliCPUToShares(mCpu)))
+			c.SetCPUShares(int64(cache.MilliCPUToShares(int64(mCpu))))
 		}
 	}
 	if p.bpoptions.PinMemory == nil || *p.bpoptions.PinMemory {
